Create folder structure: - Create Readme.md file to explain each folder. - Refactor HelloComponent
