+++ src/Readme.md
@@ -1,0 +1,24 @@
+src
+============================
+
+We will have all react source on this folder.
+
+## Structure: 
+
+```
+src/
+  ├── common/                          # common source.
+  |    └── ...
+  ├── pages/                           # page application.
+  |    └── ...
+  ├── reducers/                        # redux reducers.
+  |    └── ...
+  ├── rest-api/                        # rest-api calls.
+  |    └── ...
+  ├── app.js                           # React app wrapp component.
+  ├── index.html
+  ├── index.js                         # Init react app.
+  ├── route.js                         # Define react routes.
+  ├── store.js                         # Define react redux store.
+  └── style.scss                       # Main style sass file.
+```
