+++ src/common/Readme.md
@@ -1,0 +1,12 @@
+src/common
+============================
+
+We will add all the common source (components, constants, ...) to this folder.
+
+## Structure: 
+
+```
+src/pages                   
+    ├── components/                 # all generic components.
+    └── constants/                  # all constants.
+```
