+++ src/common/components/Hello/Hello.scss
@@ -1,0 +1,3 @@
+.HelloComponent {
+    color:red;
+}
