+++ src/common/components/Hello/HelloComponent.js
@@ -1,12 +1,10 @@
 import * as React from 'react';
-// Import css
-import css from './styles/style.scss';
 
 export const HelloComponent = () => {
     return (
-        <div>
+        <div className="HelloComponent">
             <div>
-                <h2 className="prueba">Welcome to React</h2>
+                <h2>Welcome to React</h2>
             </div>
             <p>
                 To get started, edit <code>src/HelloComponent.js</code> and save to reload.
