+++ src/common/components/Hello/package.json
@@ -1,0 +1,6 @@
+{
+    "name": "HelloComponent",
+    "version": "1.0.0",
+    "private": true,
+    "main": "./HelloComponent.js"
+}
