+++ src/common/components/Readme.md
@@ -1,0 +1,4 @@
+src/common/components
+============================
+
+We will add all the common components to this folder.
