+++ src/common/components/components.scss
@@ -1,0 +1,1 @@
+@import 'Hello/Hello.scss';
