+++ src/common/constants/Readme.md
@@ -1,0 +1,4 @@
+src/common/constants
+============================
+
+We will add all the constants to this folder.
