+++ src/pages/Readme.md
@@ -1,0 +1,16 @@
+src/pages
+============================
+
+We will have a folder for each page that we need to create.
+
+## Structure: 
+
+```
+src/pages
+    ├── module1/                         # page name.
+    |    ├── components/                 # all specific components for this page.
+    |    ├── page.js                     # page view component.
+    |    └── pageContainer.js            # redux container to map the state.
+    ├── module2/
+    └── ../
+```
