+++ src/reducers/Readme.md
@@ -1,0 +1,14 @@
+src/reducers
+============================
+
+We will have a folder for each reducer.
+
+## Structure: 
+
+```
+src/reducers
+    ├── reducer1/                       # reducer 1.
+    ├── reducer2/                       # reducer 2.
+    ├── .../
+    └── index.js                        # combine all reducers.
+```
