+++ src/rest-api/Readme.md
@@ -1,0 +1,13 @@
+src/rest-api
+============================
+
+We will have a folder for each type of content in order to manintenance all api calls.
+
+## Structure: 
+
+```
+src/rest-api                   
+    ├── module1/                 # one module for each type of content.
+    ├── module2/                 
+    └── .../                     
+```
