--- src/styles/style.scss
@@ -1,1 +1,0 @@
-@import 'reset.scss';
